refactor(webapp): move regenerate dialog logic to a hook
The dialog is now render-only; the drop handling, flags, regeneration
and dismissal guard live in useRegenerateMetadata, following the shape
of useMaintenanceCleanup.

Addresses review feedback on #409.
